pyproject.toml
[project]
name = "gem-hub-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.11",
]
